requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
faiss-cpu==1.7.4
filelock==3.18.0
Flask==2.3.3
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.25.2
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.24.3
ollama==0.1.6
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.1.2
tqdm==4.66.2
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
